+++ audiolazy/lazy_poly.py
@@ -17,21 +17,23 @@
 # Created on Sun Oct 07 2012
 # danilo [dot] bellini [at] gmail [dot] com
 """
-Polynomial model
+Polynomial model and Waring-Lagrange polynomial interpolator
 """
 
+from __future__ import division
+
 import operator
-from collections import Iterable
+from collections import Iterable, deque
 from functools import reduce
 import itertools as it
 
 # Audiolazy internal imports
-from .lazy_core import AbstractOperatorOverloaderMeta
+from .lazy_core import AbstractOperatorOverloaderMeta, StrategyDict
 from .lazy_misc import (multiplication_formatter, pair_strings_sum_formatter,
-                        meta, iteritems, xrange)
-from .lazy_stream import Stream
+                        meta, iteritems, xrange, xzip, rint)
+from .lazy_stream import Stream, tostream
 
-__all__ = ["PolyMeta", "Poly"]
+__all__ = ["PolyMeta", "Poly", "x", "lagrange", "resample"]
 
 
 class PolyMeta(AbstractOperatorOverloaderMeta):
@@ -59,12 +61,33 @@ class PolyMeta(AbstractOperatorOverloaderMeta):
 
 class Poly(meta(metaclass=PolyMeta)):
   """
-  Model for a polynomial.
+  Model for a polynomial, Laurent polynomial or a sum of powers.
 
   That's not a dict and not a list but behaves like something in between.
   The "values" method allows casting to list with list(Poly.values())
   The "terms" method allows casting to dict with dict(Poly.terms()), and give
   the terms sorted by their power value.
+
+  You can use the ``x`` object to create your own instances.
+
+  Examples
+  --------
+  >>> x ** 5 - x + 7
+  7 - x + x^5
+  >>> type(x + 1)
+  <class 'audiolazy.lazy_poly.Poly'>
+  >>> (x + 2)(17)
+  19
+  >>> (x ** 2 + 2 * x + 1)(2)
+  9
+  >>> (x ** 2 + 2 * x + 1)(.5)
+  2.25
+  >>> (x ** -2 + x)(10)
+  10.01
+  >>> spow = x ** -2.1 + x ** .3 + x ** -1 + x - 6
+  >>> value = spow(5)
+  >>> "{:.6f}".format(value) # Just to see the first few digits
+  '0.854710'
 
   """
   def __init__(self, data=None, zero=0):
@@ -155,17 +178,17 @@ class Poly(meta(metaclass=PolyMeta)):
     When value is a number, a Horner-like scheme is done.
     """
     if isinstance(value, Poly):
-     return Poly(sum(coeff * value ** power
-                     for power, coeff in iteritems(self.data)))
+      return Poly(sum(coeff * value ** power
+                      for power, coeff in iteritems(self.data)))
     if not self.data:
       return 0
+    if value == 0:
+      return self[0]
 
-    last_power = 0
-    res = 0
-    for power, coeff in sorted(iteritems(self.data), reverse=True):
-      res = coeff + res * value ** (last_power - power)
-      last_power = power
-    return res * value ** last_power
+    return reduce(
+      lambda old, new: (new[0], new[1] + old[1] * value ** (old[0] - new[0])),
+      sorted(iteritems(self.data), reverse=True) + [(0, 0)]
+    )[1]
 
   def __getitem__(self, item):
     if item in self.data:
@@ -266,3 +289,117 @@ class Poly(meta(metaclass=PolyMeta)):
     """
     import numpy as np
     return np.roots(list(self.values())[::-1]).tolist()
+
+
+x = Poly({1: 1})
+lagrange = StrategyDict("lagrange")
+
+
+@lagrange.strategy("func")
+def lagrange(pairs):
+  """
+  Waring-Lagrange interpolator function.
+
+  Parameters
+  ----------
+  pairs :
+    Iterable with pairs (tuples with two values), corresponding to points
+    ``(x, y)`` of the function.
+
+  Returns
+  -------
+  A function that returns the interpolator result for a given ``x``.
+
+  """
+  prod = lambda args: reduce(operator.mul, args)
+  xv, yv = xzip(*pairs)
+  return lambda k: sum( yv[j] * prod( (k - rk) / (rj - rk)
+                                      for rk in xv if rj != rk )
+                        for j, rj in enumerate(xv) )
+
+
+@lagrange.strategy("poly")
+def lagrange(pairs):
+  """
+  Waring-Lagrange interpolator polynomial.
+
+  Parameters
+  ----------
+  pairs :
+    Iterable with pairs (tuples with two values), corresponding to points
+    ``(x, y)`` of the function.
+
+  Returns
+  -------
+  A Poly instance that allows finding the interpolated value for any ``x``.
+
+  """
+  return lagrange.func(pairs)(x)
+
+
+@tostream
+def resample(sig, old=1, new=1, order=3, zero=0.):
+  """
+  Generic resampler based on Waring-Lagrange interpolators.
+
+  Parameters
+  ----------
+  sig :
+    Input signal (any iterable).
+  old :
+    Time duration reference (defaults to 1, allowing percentages to the ``new``
+    keyword argument). This can be float number, or perhaps a Stream instance.
+  new :
+    Time duration that the reference will have after resampling.
+    For example, if ``old = 1, new = 2``, then
+    there will be 2 samples yielded for each sample from input.
+    This can be a float number, or perhaps a Stream instance.
+  order :
+    Lagrange interpolator order. The amount of neighboring samples to be used by
+    the interpolator is ``order + 1``.
+  zero :
+    The input should be thought as zero-padded from the left with this value.
+
+  Returns
+  -------
+  The first value will be the first sample from ``sig``, and then the
+  interpolator will find the next samples towards the end of the ``sig``.
+  The actual sampling interval (or time step) for this interpolator obeys to
+  the ``old / new`` relationship.
+
+  Hint
+  ----
+  The time step can also be time-varying, although that's certainly difficult
+  to synchonize (one sample is needed for each output sample). Perhaps the
+  best approach for this case would be a ControlStream keeping the desired
+  value at any time.
+
+  Note
+  ----
+  The input isn't zero-padded at right. It means that the last output will be
+  one with interpolated with known data. For endless inputs that's ok, this
+  makes no difference, but for finite inputs that may be undesirable.
+
+  """
+  sig = Stream(sig)
+  threshold = .5 * (order + 1)
+  step = old / new
+  data = deque([zero] * (order + 1), maxlen=order + 1)
+  data.extend(sig.take(rint(threshold)))
+  idx = int(threshold)
+  isig = iter(sig)
+  if isinstance(step, Iterable):
+    step = iter(step)
+    while True:
+      yield lagrange(enumerate(data))(idx)
+      idx += next(step)
+      while idx > threshold:
+        data.append(next(isig))
+        idx -= 1
+  else:
+    while True:
+      yield lagrange(enumerate(data))(idx)
+      idx += step
+      while idx > threshold:
+        data.append(next(isig))
+        idx -= 1